requirements.txt
opensearch-py
opensearch-py-ml
python-dotenv
pandas==2.0.3
deprecated
